# openapi.yaml
openapi: 3.1.0
info:
  title: Walker&Weights API
  description: Walker&Weights API
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/api'
    description: Local server
  - url: 'https://walkerweights.vercel.app/api'
    description: Production server
...
paths:
  /admin/create:
    post:
      tags:
        - admin
      summary: Create a new user
      description: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: email of the user
                first_name:
                  type: string
                  description: first name of the user
                last_name:
                  type: string
                  description: last name of the user
                access_list:
                  type: array
                  items:
                    type: string
                  description: list of access levels for the user
                accessid:
                  type: string
                  description: objectid of the user creating
              $ref: '/models/adminSchema.js'
              required:
                - email
                - first_name
                - last_name
                - access_list
                - accessid
      responses:
        201:
          description: User created
        400:
          description: Bad request, wrong method used
        401:
          description: Bad request, missing fields
        403:
          description: Unauthorized
        500:
          description: Internal server error
...
  /admin/password-email:
    post:
      tags:
        - admin
      summary: Send password setup email
      description: Send password setup email to admin
      operationId: sendPasswordEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: email of the user
                accessid:
                  type: string
                  description: objectid of the user sending email
                first_name:
                  type: string
                  description: first name of the user
                last_name:
                  type: string
                  description: last name of the user
              required:
                - email
      responses:
        200:
          description: Email sent
        400:
          description: Bad request, wrong method used
        401:
          description: Bad request, missing fields
        403:
          description: Unauthorized
        500:
          description: Internal server error
...
  /admin/data/change-state:
    post:
      tags:
        - admin
      summary: Change company state
      description: Change company state
      operationId: changeCompanyState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accessid:
                  type: string
                  description: objectid of the user changing company state
                clientid:
                  type: string
                  description: objectid of the company
                status:
                  type: string
                  description: new state of the company
                type:
                  type: string
                  description: new type of company actions
                addTasks:
                  type: boolean
                  description: add tasks associated with new type for company
              required:
                - accessid
                - clientid
                - status
                - type
      responses:
        200:
          description: Company state changed
        400:
          description: Bad request, wrong method used
        401:
          description: Bad request, missing fields
        403:
          description: Unauthorized
        404:
          description: Company not found
        500:
          description: Internal server error
...
  /admin/data/task/send:
    post:
      tags:
        - admin
      summary: Send task
      description: Send task
      operationId: sendTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accessid:
                  type: string
                  description: objectid of the user sending task
                clientid:
                  type: string
                  description: objectid of the company
                title:
                  type: string
                  description: title of the task
                description:
                  type: string
                  description: description of the task
                type:
                  type: string
                  description: type of the task, either form or file format (pdf, xlsx)
                formID:
                  type: string
                  description: objectid of the form; required if type is form
              required:
                - accessid
                - clientid
                - title
                - description
                - type
      responses:
        201:
          description: Task sent
        400:
          description: Bad request, wrong method used
        401:
          description: Bad request, missing fields
        403:
          description: Unauthorized
        404:
          description: Company not found
        500:
          description: Internal server error
...